+++ packages/number-lookup/CHANGELOG.md
@@ -1,2 +1,5 @@
+## Version 1.4.1
+- [Bugfix] Make the lazyClient public in the NumberLookupService to be able to override it with a custom one.
+
 ## Version 1.4.0
 - Initial version